+++ packages/react/CHANGELOG.md
@@ -28,4 +28,4 @@
 
 ### Patch Changes
 
-- c81587f: use-search-params-state mvp release
+- c81587f: sp-hooks mvp release